Update patch set 8
Patch Set 8:

Seems odd to me that this is needed. What is 1.34 even doing under the hood for this to be required? Weird. But if it fixes the issue, I'm more than happy to +2 this because obviously we'll want this to work under 1.34 as well.

Patch-set: 8
Reviewer: Jack Phoenix <243@e9e9afe9-4712-486d-8885-f54b72dd1951>