Bugfix for Special:MassRating when user's UI language is not site's content language: get ratings JSON in content language (only)
This lets all users regardless of their chosen language to use the special page in their own language.

The actual rating data, as defined in [[MediaWiki:Are-ratings]], is untranslatable currently; this was the case before and still is.
But at least now Special:MassRatings shows something (when said MW message has some content), even for international users.

Change-Id: I2f700fbfaed7245e40c9355dc851004d553e9acb