Rating needs to be seeded with a (string) codename, not with a Rating object
My bad.

This fixes the Special:Log/ratings entries, which previously were lacking the default rating name and showing up as such:
10:33, 3 May 2024 Ashley changed the rating of Test Page 666 from to Featured ‎ (Great page with lots of useful content, love it.)

Now they show up like this if and when a page has no prior ratings:
10:33, 3 May 2024 Ashley changed the rating of Test Page 666 from Standard to Featured ‎ (Great page with lots of useful content, love it.)

(This is, of course, assuming that the default rating is called "Standard", which it is, if you copied your ARE configuration from Brickipedia.)

Follow-up to d44ca763fc6518bbaf7a02550aded55ba780e06b.

Change-Id: I05f616df8471ae901b2fdf8045aa7d592c032f9e