Move PHP source files to a src/ subdirectory and its subdirectories, update paths accordingly
This is the de facto standard these days (I prefer "src" over "includes" because you can tab-complete easily by hitting the S key + tab, whereas with the I key + tab you'll be suggested "i18n" most of the time).
Also, having a neat directory structure is nice and this way you don't have to wonder where all the files are.

Change-Id: Ie709b6bc6752d94ee479267d524285b121fd8349
TODO: In the future it'd be nice to namespace the extension and use AutoloadNamespaces etc. but that can wait.