New API query module for listing pages' ARE ratings
Named so because I sporked this from an existing-but-uncommitted thing I wrote for the VoteNY and I don't want them to conflict...but neither extension is yet namespaced.
So tacking in an "ARE" into the class/file name here was the easiest thing to do, and for the api.php action name as well.

Tested, but not a lot, since I only have a handful of ARE-rated pages in my dev box.

Bug: T403131
Change-Id: I8e4b73ebadd2e53fdca714473f9e76e3a98fbf45